backend/requirements.txt
alembic==1.6.5
amqp==2.6.1
arrow==1.1.0
asgiref==3.3.4
asyncpg==0.23.0
billiard==3.6.4.0
binaryornot==0.4.4
cachetools==4.2.2
celery==4.4.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cookiecutter==1.7.3
cssselect==1.1.0
cssutils==2.3.0
databases[postgresql]==0.4.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.6
fastapi==0.65.2
FastAPI-SQLAlchemy==0.2.1
flower==0.9.7
greenlet==1.1.0
h11==0.12.0
humanize==3.9.0
idna==2.10
importlib-metadata==4.5.0
Jinja2==3.0.1
jinja2-time==0.2.0
jose==1.0.0
kombu==4.6.11
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
plotly==5.0.0
postgres==3.0.0
poyo==0.5.0
premailer==3.9.0
prometheus-client==0.8.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pyasn1==0.4.8
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.18
starlette==0.14.2
tenacity==7.0.0
text-unidecode==1.3
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
vine==1.3.0
zipp==3.4.1
